requirements.txt
detectron2==0.6
fiftyone==0.20.0
fiftyone_db==0.4.0
hydra-core==1.3.2
imgaug==0.4.0
numpy==1.24.2
opencv_python_headless==4.7.0.72
ultralytics==8.0.58
